+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
 *    Adding timestamps to the output names and append the new outputs to the existing json file
 *    Create functionality for users to rename outputs
 *    Create functionality to add comments to outputs
+*    Create functionality to add currently unsupported outputs to the list of outputs
 
 ## Version 0.0.5 (Nov 02, 2022)